chapter-4/extract-metadata-from-files/requirements.txt
ffmpeg-python
Pillow
pikepdf
tinytag